fix eror

+++ src/main/java/com/example/backend/dto/AdminProductResponse.java
@@ -4,14 +4,16 @@ import com.example.backend.entity.Category;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class AdminProductResponse {
     private Long id;
     private String name;
     private String brand;
-    private double price;
-    private double discountPrice;
+    private BigDecimal price;
+    private BigDecimal discountPrice;
     private int stock;
     private boolean active;
     private Category category;
